FEAT: add use context for data and migrate functions
