+++ src/App.jsx
@@ -1,6 +1,7 @@
 import { Suspense } from "react";
 import { Route, Routes } from "react-router-dom";
 import "./App.css";
+import { DataProvider } from "./context/dataContext";
 import { NoRequireAuth } from "./services/NoRequireAuth";
 import { RequireAuth } from "./services/RequireAuth";
 import { DashboardView, EditProfile, Profile } from "./views/dashboard";
@@ -55,9 +56,11 @@ function App() {
         path="/dashboard"
         element={
           <Suspense fallback={<h1>loading...</h1>}>
-            <RequireAuth>
-              <DashboardView />
-            </RequireAuth>
+            <DataProvider>
+              <RequireAuth>
+                <DashboardView />
+              </RequireAuth>
+            </DataProvider>
           </Suspense>
         }
       >
