+++ src/index.css
@@ -16,6 +16,10 @@ code {
   width: 0;
   height: 0;
 }
+.withScrollbar::-webkit-scrollbar {
+  width: 0.3em;
+  width: 0.3em;
+}
 
 *::-webkit-scrollbar-track {
   background-color: rgba(255, 255, 255, 0);
