+++ src/services/RequireAuth.jsx
@@ -7,15 +7,11 @@ export function RequireAuth({ children }) {
   const auth = useAuth();
   const toast = useToast();
   let location = useLocation();
-  console.log(auth.isLoadingView);
   if (
     !auth.user &&
     !auth.isLoadingView &&
     location.pathname !== "/login/professional"
   ) {
-    console.log("no hay user");
-    console.log(auth.isLoadingView);
-
     toast({
       title: "Login required",
       description: "You must be logged in to view this page",
