+++ src/services/sessions/category-services.jsx
@@ -10,7 +10,6 @@ export async function indexCategories() {
     },
   };
   const response = await axios.request(options);
-  console.log(response.data);
   // const { token, ...user } = response.data;
 
   return response.data;
